Implemented logic to ensure only task authors can access, update, or delete their tasks

+++ backend/src/task/task.service.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Injectable, InternalServerErrorException, Logger, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, InternalServerErrorException, Logger, NotFoundException, Request, UnauthorizedException } from '@nestjs/common';
 import { TaskService } from './task.interface';
 import { Task } from './task.schema';
 import { InjectModel } from '@nestjs/mongoose';
@@ -15,11 +15,19 @@ export class TaskServiceImp implements TaskService {
                 ) {}
    
 
-   async delete(id: string): Promise<void> {
+   async delete(id: string ,req ): Promise<void> {
+       
+        const authenticatedUser = req.user
+        if(authenticatedUser == null) throw new BadRequestException("bad request")
+
+
         const taskObjectId = generateObjectId(id)
         const task = await this.findById(id) 
         if(task == null) throw new NotFoundException("task is not found")
-        
+        const userTaskId = task.user 
+        console.log(userTaskId , "userTaskId")
+        if(userTaskId !== authenticatedUser._id) throw new UnauthorizedException("user is not authorized to delete this task")
+
         try{
             await this.taskModel.findByIdAndDelete(taskObjectId)
         }catch(e){
@@ -28,13 +36,22 @@ export class TaskServiceImp implements TaskService {
         }
     }
 
-    async update(id: string, updatedTaskDto: UpdatedTaskDto): Promise<Task | null> {
+    async update(id: string, updatedTaskDto: UpdatedTaskDto , req ): Promise<Task | null> {
+        
+        const authenticatedUser = req.user
+        if(authenticatedUser == null) throw new BadRequestException("bad request")
+
         const taskObjectId : Types.ObjectId = generateObjectId(id)
         if(updatedTaskDto.completed != null && !updatedTaskDto.completed ) {
             throw new BadRequestException("you cannot undo the completed status of the task")
         }
         const task = await this.findById(id)
         if(task == null) throw new NotFoundException("task is not found")
+        
+        const userTaskId = task.user 
+        console.log(userTaskId , "userTaskId")
+        if(userTaskId !== authenticatedUser._id) throw new UnauthorizedException("user is not authorized to update this task")
+
         try{
             return await this.taskModel.findByIdAndUpdate(taskObjectId , updatedTaskDto , {new : true})
         }catch(e) {
@@ -43,13 +60,13 @@ export class TaskServiceImp implements TaskService {
         }    
     }
    
-    async findById(id: string): Promise<Task | null> {
+   private async findById(id: string): Promise<Task | null> {
         const taskObjectId = generateObjectId(id)
         const task = await this.taskModel.findById(taskObjectId)
         return task
     }
 
-    async findAllByUserId(userId : string): Promise<Task[]> {
+    async findAllByUserId(userId : string ): Promise<Task[]> {
         const userObjectId = generateObjectId(userId)
         const tasks = await this.taskModel.find({user : userObjectId})
         return tasks
